+++ openeogeotrellis/backend.py
@@ -21,14 +21,12 @@ from subprocess import CalledProcessError
 from typing import Callable, Dict, Iterable, List, Optional, Tuple, Union
 from urllib.parse import urlparse
 
-import dateutil
 import flask
 import geopyspark as gps
 import kazoo.exceptions
 import openeo.udf
 import pkg_resources
 import pystac
-import pystac_client
 import requests
 import shapely.geometry.base
 from deprecated import deprecated
@@ -36,24 +34,24 @@ from geopyspark import LayerType, Pyramid, TiledRasterLayer
 from openeo.internal.process_graph_visitor import ProcessGraphVisitor
 from openeo.metadata import Band, BandDimension, Dimension, SpatialDimension, TemporalDimension
 from openeo.util import TimingLogger, deep_get, dict_no_none, repr_truncate, rfc3339, str_truncate
-from openeo_driver import backend, filter_properties
+from openeo_driver import backend
 from openeo_driver.backend import BatchJobMetadata, ErrorSummary, LoadParameters, OidcProvider, ServiceMetadata
 from openeo_driver.config.load import ConfigGetter
 from openeo_driver.datacube import DriverDataCube, DriverVectorCube
 from openeo_driver.datastructs import SarBackscatterArgs
 from openeo_driver.delayed_vector import DelayedVector
 from openeo_driver.dry_run import SourceConstraint
-from openeo_driver.errors import (InternalException, JobNotFinishedException, JobNotFoundException, OpenEOApiException,
-                                  ProcessParameterUnsupportedException, ServiceUnsupportedException,
+from openeo_driver.errors import (InternalException, JobNotFinishedException, OpenEOApiException,
+                                  ServiceUnsupportedException,
                                   ProcessParameterInvalidException, )
 from openeo_driver.jobregistry import (DEPENDENCY_STATUS, JOB_STATUS, ElasticJobRegistry, PARTIAL_JOB_STATUS,
                                        get_ejr_credentials_from_env)
 from openeo_driver.ProcessGraphDeserializer import ENV_SAVE_RESULT, ConcreteProcessing
 from openeo_driver.save_result import ImageCollectionResult
 from openeo_driver.users import User
-from openeo_driver.util.geometry import BoundingBox, GeometryBufferer
+from openeo_driver.util.geometry import BoundingBox
 from openeo_driver.util.http import requests_with_retry
-from openeo_driver.util.utm import area_in_square_meters, utm_zone_from_epsg
+from openeo_driver.util.utm import area_in_square_meters
 from openeo_driver.utils import EvalEnv, generate_unique_id, to_hashable, smart_bool
 from pandas import Timedelta
 from py4j.java_gateway import JavaObject, JVMView
@@ -65,11 +63,11 @@ from urllib3 import Retry
 from xarray import DataArray
 import numpy as np
 
-from openeogeotrellis import sentinel_hub
+from openeogeotrellis import sentinel_hub, load_stac
 from openeogeotrellis.config import get_backend_config
 from openeogeotrellis.configparams import ConfigParams
 from openeogeotrellis.geopysparkdatacube import GeopysparkCubeMetadata, GeopysparkDataCube
-from openeogeotrellis.integrations.etl_api import EtlApi, get_etl_api, get_etl_api_credentials_from_env
+from openeogeotrellis.integrations.etl_api import get_etl_api
 from openeogeotrellis.integrations.hadoop import setup_kerberos_auth
 from openeogeotrellis.integrations.kubernetes import k8s_job_name, kube_client, truncate_job_id_k8s, k8s_render_manifest_template
 from openeogeotrellis.integrations.traefik import Traefik
@@ -759,434 +757,34 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
         return cube
 
     def load_stac(self, url: str, load_params: LoadParameters, env: EvalEnv) -> GeopysparkDataCube:
-        logger.info("load_stac from url {u!r} with load params {p!r}".format(u=url, p=load_params))
-
-        no_data_available_exception = OpenEOApiException(message="There is no data available for the given extents.",
-                                                         code="NoDataAvailable", status_code=400)
-        properties_unsupported_exception = ProcessParameterUnsupportedException("load_stac", "properties")
-
-        user: Union[User, None] = env["user"]
-
-        requested_bbox = BoundingBox.from_dict_or_none(
-            load_params.spatial_extent, default_crs="EPSG:4326"
-        )
-
-        temporal_extent = load_params.temporal_extent
-        from_date, until_date = map(dt.datetime.fromisoformat, normalize_temporal_extent(temporal_extent))
-        to_date = (dt.datetime.combine(until_date, dt.time.max, until_date.tzinfo) if from_date == until_date
-                   else until_date - dt.timedelta(milliseconds=1))
-
-        def intersects_spatiotemporally(itm: pystac.Item) -> bool:
-            def intersects_temporally() -> bool:
-                nominal_date = itm.datetime or dateutil.parser.parse(itm.properties["start_datetime"])
-                return from_date <= nominal_date <= to_date
-
-            def intersects_spatially() -> bool:
-                if not requested_bbox or itm.bbox is None:
-                    return True
-
-                requested_bbox_lonlat = requested_bbox.reproject("EPSG:4326")
-                return requested_bbox_lonlat.as_polygon().intersects(
-                    Polygon.from_bounds(*itm.bbox)
-                )
-
-            return intersects_temporally() and intersects_spatially()
-
-        def supports_item_search(coll: pystac.Collection) -> bool:
-            # TODO: use pystac_client instead?
-            conforms_to = coll.get_root().extra_fields.get("conformsTo", [])
-            return any(conformance_class.endswith("/item-search") for conformance_class in conforms_to)
-
-        def is_band_asset(asset: pystac.Asset) -> bool:
-            return "eo:bands" in asset.extra_fields
-
-        def get_band_names(itm: pystac.Item, asst: pystac.Asset) -> List[str]:
-            def get_band_name(eo_band) -> str:
-                if isinstance(eo_band, dict):
-                    return eo_band["name"]
-
-                # can also be an index into a list of bands elsewhere.
-                # TODO: still necessary to support this? See https://github.com/Open-EO/openeo-geopyspark-driver/issues/619
-                assert isinstance(eo_band, int)
-                eo_band_index = eo_band
-
-                eo_bands_location = (itm.properties if "eo:bands" in itm.properties
-                                     else itm.get_collection().summaries.to_dict())
-                return get_band_name(eo_bands_location["eo:bands"][eo_band_index])
-
-            return [get_band_name(eo_band) for eo_band in asst.extra_fields["eo:bands"]]
-
-        def get_proj_metadata(itm: pystac.Item, asst: pystac.Asset) -> (Optional[int],
-                                                                        Optional[Tuple[float, float, float, float]],
-                                                                        Optional[Tuple[int, int]]):
-            """Returns EPSG code, bbox (in that EPSG) and number of pixels (rows, cols), if available."""
-            epsg = asst.extra_fields.get("proj:epsg") or itm.properties.get("proj:epsg")
-            bbox = asst.extra_fields.get("proj:bbox") or itm.properties.get("proj:bbox")
-            shape = asst.extra_fields.get("proj:shape") or itm.properties.get("proj:shape")
-            return (epsg,
-                    tuple(map(float, bbox)) if bbox else None,
-                    tuple(shape) if shape else None)
-
-        def matches_metadata_properties(itm: pystac.Item) -> bool:
-            literal_matches = {property_name: filter_properties.extract_literal_match(condition)
-                               for property_name, condition in load_params.properties.items()}
-
-            def operator_value(criterion: Dict[str, object]) -> (str, object):
-                if len(criterion) != 1:
-                    raise ValueError(f'expected a single criterion, was {criterion}')
-
-                (operator, value), = criterion.items()
-                return operator, value
-
-            for property_name, criterion in literal_matches.items():
-                if property_name not in itm.properties:
-                    return False
-
-                item_value = itm.properties[property_name]
-                operator, criterion_value = operator_value(criterion)
-
-                if operator == 'eq' and item_value != criterion_value:
-                    return False
-                if operator == 'lte' and item_value is not None and item_value > criterion_value:
-                    return False
-                if operator == 'gte' and item_value is not None and item_value < criterion_value:
-                    return False
-
-            return True
-
-        # TODO: `user` might be None
-        dependency_job_info = extract_own_job_info(url, user_id=user.user_id, batch_jobs=self.batch_jobs)
-        collection = None
-        if dependency_job_info:
-            intersecting_items = []
-
-            for asset_id, asset in self.batch_jobs.get_result_assets(job_id=dependency_job_info.id,
-                                                                     user_id=user.user_id).items():
-                pystac_item = pystac.Item(id=asset_id, geometry=asset["geometry"], bbox=asset["bbox"],
-                                          datetime=rfc3339.parse_datetime(asset["datetime"], with_timezone=True),
-                                          properties={"datetime": asset["datetime"]})
-
-                if intersects_spatiotemporally(pystac_item) and "data" in asset.get("roles", []):
-                    eo_bands = [{"name": b.name} for b in asset["bands"]]
-                    pystac_asset = pystac.Asset(href=asset["href"], extra_fields={"eo:bands": eo_bands})
-                    pystac_item.add_asset(asset_id, pystac_asset)
-                    intersecting_items.append(pystac_item)
-
-            band_names = []
-        else:
-            stac_object = pystac.read_file(href=url)
-
-            if isinstance(stac_object, pystac.Item):
-                if load_params.properties:
-                    raise properties_unsupported_exception
-
-                item = stac_object
-
-                if not intersects_spatiotemporally(item):
-                    raise no_data_available_exception
-
-                if "eo:bands" in item.properties:
-                    eo_bands_location = item.properties
-                elif item.get_collection() is not None:
-                    collection = item.get_collection()
-                    eo_bands_location = item.get_collection().summaries.lists
-                else:
-                    # TODO: band order is not "stable" here, see https://github.com/Open-EO/openeo-processes/issues/488
-                    eo_bands_location = {}
-                band_names = [b["name"] for b in eo_bands_location.get("eo:bands", [])]
-
-                intersecting_items = [item]
-            elif isinstance(stac_object, pystac.Collection) and supports_item_search(stac_object):
-                collection = stac_object
-                collection_id = collection.id
-
-                root_catalog = collection.get_root()
-
-                band_names = [b["name"] for b in collection.summaries.lists.get("eo:bands", [])]
-
-                client = pystac_client.Client.open(root_catalog.get_self_href())
-                search_request = client.search(
-                    method="GET",
-                    collections=collection_id,
-                    bbox=requested_bbox.reproject("EPSG:4326").as_wsen_tuple() if requested_bbox else None,
-                    limit=20,
-                    datetime=f"{from_date.isoformat().replace('+00:00', 'Z')}/{to_date.isoformat().replace('+00:00', 'Z')}",  # inclusive
-                )
-
-                logger.info(f"STAC API request: GET {search_request.url_with_parameters()}")
-
-                # TODO: use server-side filtering as well (at least STAC API Filter Extension)
-                intersecting_items = filter(lambda itm: matches_metadata_properties(itm), search_request.items())
-            else:
-                assert isinstance(stac_object, pystac.Catalog)  # static Catalog + Collection
-                catalog = stac_object
-
-                if load_params.properties:
-                    raise properties_unsupported_exception
-
-                if isinstance(catalog, pystac.Collection):
-                    collection = catalog
-
-                band_names = [b["name"] for b in (catalog.summaries.lists if isinstance(catalog, pystac.Collection)
-                                                  else catalog.extra_fields.get("summaries", {})).get("eo:bands", [])]
-
-                def intersecting_catalogs(root: pystac.Catalog) -> Iterable[pystac.Catalog]:
-                    def intersects_spatiotemporally(coll: pystac.Collection) -> bool:
-                        def intersects_spatially(bbox) -> bool:
-                            if not requested_bbox:
-                                return True
-
-                            requested_bbox_lonlat = requested_bbox.reproject("EPSG:4326")
-                            return requested_bbox_lonlat.as_polygon().intersects(
-                                Polygon.from_bounds(*bbox)
-                            )
-
-                        def intersects_temporally(interval) -> bool:
-                            start, end = interval
-
-                            if start is not None and end is not None:
-                                return to_date >= start and from_date <= end
-                            if start is not None:
-                                return to_date >= start
-                            if end is not None:
-                                return from_date <= end
-                            return True
-
-                        bboxes = coll.extent.spatial.bboxes
-                        intervals = coll.extent.temporal.intervals
-
-                        if len(bboxes) > 1 and not any(intersects_spatially(bbox) for bbox in bboxes[1:]):
-                            return False
-                        if len(bboxes) == 1 and not intersects_spatially(bboxes[0]):
-                            return False
-
-                        if len(intervals) > 1 and not any(intersects_temporally(interval)
-                                                          for interval in intervals[1:]):
-                            return False
-                        if len(intervals) == 1 and not intersects_temporally(intervals[0]):
-                            return False
-
-                        return True
-
-                    if isinstance(root, pystac.Collection) and not intersects_spatiotemporally(root):
-                        return []
-
-                    yield root
-                    for child in root.get_children():
-                        yield from intersecting_catalogs(child)
-
-                intersecting_items = (itm
-                                      for intersecting_catalog in intersecting_catalogs(root=catalog)
-                                      for itm in intersecting_catalog.get_items() if intersects_spatiotemporally(itm))
-
-        jvm = get_jvm()
-
-        opensearch_client = jvm.org.openeo.geotrellis.file.FixedFeaturesOpenSearchClient()
-
-        stac_bbox = None
-        items_found = False
-        proj_epsg = None
-        proj_bbox = None
-        proj_shape = None
-
-        netcdf_with_time_dimension = False
-        if collection is not None:
-            #we found some collection level metadata
-            item_assets = collection.extra_fields.get("item_assets", {})
-            dimensions = set([tuple(v.get("dimensions")) for i in item_assets.values() if "cube:variables" in i for v in i.get("cube:variables",{}).values() ])
-            #this is one way to determine if a time dimension is used, but it does depend on the use of item_assets and datacube extension.
-            netcdf_with_time_dimension = len(dimensions) == 1 and "time" in dimensions.pop()
-
-        for itm in intersecting_items:
-            band_assets = {asset_id: asset for asset_id, asset
-                           in dict(sorted(itm.get_assets().items())).items() if is_band_asset(asset)}
-
-            builder = jvm.org.openeo.opensearch.OpenSearchResponses.featureBuilder()
-
-            builder = builder.withId(itm.id).withBBox( itm.bbox[0],itm.bbox[1],itm.bbox[2],itm.bbox[3]) \
-                .withNominalDate(itm.properties.get("datetime") or itm.properties["start_datetime"])
-
-            for asset_id, asset in band_assets.items():
-                asset_band_names = get_band_names(itm, asset)
-                for asset_band_name in asset_band_names:
-                    if asset_band_name not in band_names:
-                        band_names.append(asset_band_name)
-
-                proj_epsg, proj_bbox, proj_shape = get_proj_metadata(itm, asset)
-
-                builder = builder.addLink(asset.get_absolute_href() or asset.href, asset_id, asset_band_names)
-
-            if proj_epsg:
-                builder = builder.withCRS(f"EPSG:{proj_epsg}")
-            if proj_bbox:
-                builder = builder.withRasterExtent(*proj_bbox)
-
-            if proj_bbox and proj_shape:
-                cell_width, cell_height = self.compute_cellsize(proj_bbox, proj_shape)
-                builder = builder.withResolution(cell_width)
-
-            f = builder.build()
-            opensearch_client.addFeature(f)
-
-            item_bbox = BoundingBox.from_wsen_tuple(
-                itm.bbox, crs="EPSG:4326"
-            )
-
-            stac_bbox = (item_bbox if stac_bbox is None
-                         else BoundingBox.from_wsen_tuple(item_bbox.as_polygon().union(stac_bbox.as_polygon()).bounds,
-                                                          stac_bbox.crs))
-
-            items_found = True
-
-        if not items_found:
-            raise no_data_available_exception
-
-        if not band_names:
-            raise OpenEOApiException(
-                message=f'No band assets found in items; a band asset requires an "eo:bands" property with a "name".',
-                status_code=400)
-
-        target_bbox = requested_bbox or stac_bbox
-
-        if proj_epsg and proj_bbox and proj_shape:  # exact resolution
-            target_epsg = proj_epsg
-            cell_width, cell_height = self.compute_cellsize(proj_bbox, proj_shape)
-        elif proj_epsg:  # about 10m in given CRS
-            target_epsg = proj_epsg
-            try:
-                utm_zone_from_epsg(proj_epsg)
-                cell_width = cell_height = 10.0
-            except ValueError:
-                cell_width = cell_height = GeometryBufferer.transform_meter_to_crs(
-                    10.0, f"EPSG:{proj_epsg}", loi=((target_bbox.east - target_bbox.west) / 2,
-                                                    (target_bbox.north - target_bbox.south / 2)))
-        else:  # 10m UTM
-            target_epsg = target_bbox.best_utm()
-            cell_width = cell_height = 10.0
-
-        metadata = GeopysparkCubeMetadata(metadata={}, dimensions=[
-            # TODO: detect actual dimensions instead of this simple default?
-            SpatialDimension(name="x", extent=[]), SpatialDimension(name="y", extent=[]),
-            TemporalDimension(name='t', extent=[]),
-            BandDimension(name="bands", bands=[Band(band_name) for band_name in band_names])
-        ])
-
-        if load_params.bands:
-            metadata = metadata.filter_bands(load_params.bands)
-
-        band_names = metadata.band_names
-
-        if netcdf_with_time_dimension:
-            pyramid_factory = jvm.org.openeo.geotrellis.layers.NetCDFCollection
-        else:
-            pyramid_factory = jvm.org.openeo.geotrellis.file.PyramidFactory(
-                opensearch_client,
-                url,  # openSearchCollectionId, not important
-                band_names,  # openSearchLinkTitles
-                None,  # rootPath, not important
-                jvm.geotrellis.raster.CellSize(cell_width, cell_height),
-                False  # experimental
-            )
-
-        extent = jvm.geotrellis.vector.Extent(*map(float, target_bbox.as_wsen_tuple()))
-        extent_crs = target_bbox.crs
-
-        geometries = load_params.aggregate_spatial_geometries
-
-        if not geometries:
-            projected_polygons = jvm.org.openeo.geotrellis.ProjectedPolygons.fromExtent(extent, extent_crs)
-        else:
-            projected_polygons = to_projected_polygons(
-                jvm, geometries, crs=extent_crs, buffer_points=True
-            )
-
-        projected_polygons = getattr(
-            getattr(jvm.org.openeo.geotrellis, "ProjectedPolygons$"), "MODULE$"
-        ).reproject(projected_polygons, target_epsg)
-
-        metadata_properties = {}
-        correlation_id = env.get('correlation_id', '')
-
-        data_cube_parameters = jvm.org.openeo.geotrelliscommon.DataCubeParameters()
-        getattr(data_cube_parameters, "layoutScheme_$eq")("FloatingLayoutScheme")
-
-        feature_flags = load_params.get("featureflags", {})
-        tilesize = feature_flags.get("tilesize", None)
-        if tilesize:
-            getattr(data_cube_parameters, "tileSize_$eq")(tilesize)
-        single_level = env.get('pyramid_levels', 'all') != 'all'
-
-        if netcdf_with_time_dimension:
-            pyramid = pyramid_factory.datacube_seq(projected_polygons, from_date.isoformat(), to_date.isoformat(),
-                                                   metadata_properties, correlation_id, data_cube_parameters, opensearch_client)
-        elif single_level:
-            pyramid = pyramid_factory.datacube_seq(projected_polygons, from_date.isoformat(), to_date.isoformat(),
-                                                   metadata_properties, correlation_id, data_cube_parameters)
-        else:
-            if requested_bbox:
-                extent = jvm.geotrellis.vector.Extent(*map(float, requested_bbox.as_wsen_tuple()))
-                extent_crs = requested_bbox.crs
-            else:
-                extent = jvm.geotrellis.vector.Extent(-180.0, -90.0, 180.0, 90.0)
-                extent_crs = "EPSG:4326"
-
-            pyramid = pyramid_factory.pyramid_seq(
-                extent, extent_crs, from_date.isoformat(), to_date.isoformat(),
-                metadata_properties, correlation_id
-            )
-
-        metadata = metadata.filter_temporal(from_date.isoformat(), to_date.isoformat())
-
-        metadata = metadata.filter_bbox(
-            west=extent.xmin(),
-            south=extent.ymin(),
-            east=extent.xmax(),
-            north=extent.ymax(),
-            crs=extent_crs,
-        )
-
-        temporal_tiled_raster_layer = jvm.geopyspark.geotrellis.TemporalTiledRasterLayer
-        option = jvm.scala.Option
-
-        # noinspection PyProtectedMember
-        levels = {pyramid.apply(index)._1(): TiledRasterLayer(LayerType.SPACETIME, temporal_tiled_raster_layer(
-            option.apply(pyramid.apply(index)._1()), pyramid.apply(index)._2())) for index in
-                  range(0, pyramid.size())}
-
-        return GeopysparkDataCube(pyramid=gps.Pyramid(levels), metadata=metadata)
-
-    def compute_cellsize(self, proj_bbox, proj_shape):
-        xmin, ymin, xmax, ymax = proj_bbox
-        rows, cols = proj_shape
-        cell_width = (xmax - xmin) / cols
-        cell_height = (ymax - ymin) / rows
-        return cell_width,cell_height
+        return load_stac.load_stac(url, load_params, env, layer_properties={}, batch_jobs=self.batch_jobs)
 
     def load_ml_model(self, model_id: str) -> 'JavaObject':
 
         # Trick to make sure IDE infers right type of `self.batch_jobs` and can resolve `get_job_output_dir`
         gps_batch_jobs: GpsBatchJobs = self.batch_jobs
 
-        def _create_model_dir():
+        def _create_model_dir(use_s3=False):
+            if use_s3:
+                # s3a://
+                return f"openeo-ml-models-dev/{generate_unique_id(prefix='model')}"
+
             def _set_permissions(job_dir: Path):
-                if not ConfigParams().is_kube_deploy:
-                    try:
-                        shutil.chown(job_dir, user = None, group = 'eodata')
-                    except LookupError as e:
-                        logger.warning(f"Could not change group of {job_dir} to eodata.")
+                try:
+                    shutil.chown(job_dir, user = None, group = 'eodata')
+                except LookupError as e:
+                    logger.warning(f"Could not change group of {job_dir} to eodata.")
                 add_permissions(job_dir, stat.S_ISGID | stat.S_IWGRP)  # make children inherit this group
-            ml_models_path = gps_batch_jobs.get_job_output_dir("ml_models")
-            if not os.path.exists(ml_models_path):
-                logger.info("Creating directory: {}".format(ml_models_path))
-                os.makedirs(ml_models_path)
-                _set_permissions(ml_models_path)
-            # Use a random id to avoid collisions.
-            model_dir_path = ml_models_path / generate_unique_id(prefix="model")
-            if not os.path.exists(model_dir_path):
-                logger.info("Creating directory: {}".format(model_dir_path))
-                os.makedirs(model_dir_path)
-                _set_permissions(model_dir_path)
-            return str(model_dir_path)
+
+            result_dir = gps_batch_jobs.get_job_output_dir("ml_models")
+            result_path = result_dir / generate_unique_id(prefix="model")
+            result_dir_exists = os.path.exists(result_dir)
+            logger.info("Creating directory: {}".format(result_path))
+            os.makedirs(result_path)
+            if not result_dir_exists:
+                _set_permissions(result_dir)
+            _set_permissions(result_path)
+            return str(result_path)
 
         if model_id.startswith('http'):
             # Load the model using its STAC metadata file.
@@ -1216,8 +814,32 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
             model_url = checkpoints[0]["href"]
             architecture = metadata["properties"]["ml-model:architecture"]
             # Download the model to the ml_models folder and load it as a java object.
-            model_dir_path = _create_model_dir()
+            use_s3 = ConfigParams().is_kube_deploy
+            model_dir_path = _create_model_dir(use_s3)
             if architecture == "random-forest":
+                if use_s3:
+                    with tempfile.TemporaryDirectory() as tmp_dir:
+                        # Download to tmp_dir and unpack it there.
+                        tmp_path = Path(tmp_dir + "/randomforest.model.tar.gz")
+                        logger.info(f"Downloading ml_model from {model_url} to {tmp_path}")
+                        with open(tmp_path, 'wb') as f:
+                            f.write(requests.get(model_url).content)
+                        shutil.unpack_archive(tmp_path, extract_dir = tmp_dir, format = 'gztar')
+                        # Upload the unpacked model to s3.
+                        unpacked_model_path = str(tmp_path).replace(".tar.gz", "")
+                        logger.info(f"Uploading ml_model to {model_dir_path}")
+                        path_split = model_dir_path.split("/")
+                        bucket, key = path_split[0], path_split[1]
+                        s3 = s3_client()
+                        for root, dirs, files in os.walk(unpacked_model_path):
+                            for file in files:
+                                relative_filepath = os.path.relpath(os.path.join(root, file), tmp_dir)
+                                s3.upload_file(os.path.join(root, file), bucket, key + "/" + relative_filepath)
+                        # Load the spark model using the new s3 path.
+                        s3_path = f"s3a://{model_dir_path}/randomforest.model/"
+                        logger.info("Loading ml_model using filename: {}".format(s3_path))
+                        model: JavaObject = RandomForestModel._load_java(sc = gps.get_spark_context(), path = s3_path)
+                        return model
                 dest_path = Path(model_dir_path + "/randomforest.model.tar.gz")
                 with open(dest_path, 'wb') as f:
                     f.write(requests.get(model_url).content)
@@ -1225,7 +847,18 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
                 unpacked_model_path = str(dest_path).replace(".tar.gz", "")
                 logger.info("Loading ml_model using filename: {}".format(unpacked_model_path))
                 model: JavaObject = RandomForestModel._load_java(sc=gps.get_spark_context(), path="file:" + unpacked_model_path)
+                return model
             elif architecture == "catboost":
+                if use_s3:
+                    # TODO: Verify that local files work. If it does, we can remove the model_dir_path implementation.
+                    # Download the model to the tmp directory and load it as a java object.
+                    with tempfile.TemporaryDirectory() as tmp_dir:
+                        tmp_path = Path(tmp_dir + "/catboost_model.cbm")
+                        logger.info(f"Downloading ml_model from {model_url} to {tmp_path}")
+                        with open(tmp_path, 'wb') as f:
+                            f.write(requests.get(model_url).content)
+                        model: JavaObject = CatBoostClassificationModel.load_native_model(tmp_path)
+                        return model
                 filename = Path(model_dir_path + "/catboost_model.cbm")
                 with open(filename, 'wb') as f:
                     f.write(requests.get(model_url).content)
@@ -1420,7 +1053,14 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
                 elif root_cause_message:
                     udf_stacktrace = GeoPySparkBackendImplementation.extract_udf_stacktrace(root_cause_message)
                     if udf_stacktrace:
-                        summary = f"UDF Exception during Spark execution: {udf_stacktrace}"
+                        if len(udf_stacktrace) > width - 150:
+                            udf_stacktrace_list = udf_stacktrace.split("\n")
+                            udf_stacktrace_new = "\n".join(
+                                udf_stacktrace_list[:5] + ["... skipped stack frames ..."] + udf_stacktrace_list[-5:]
+                            )
+                            if len(udf_stacktrace_new) < width - 150:
+                                udf_stacktrace = udf_stacktrace_new
+                        summary = f"UDF exception while evaluating processing graph. Please check your user defined functions. {udf_stacktrace}"
                     else:
                         summary = f"Exception during Spark execution: {root_cause_class_name}: {root_cause_message}"
                 else:
@@ -1496,11 +1136,16 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
         self.batch_jobs.set_terrascope_access_token_getter(get_terrascope_access_token)
 
     def request_costs(
-        self, *, user: Optional[User] = None, user_id: Optional[str] = None, request_id: str, success: bool
+        self,
+        *,
+        user: User,
+        job_options: Union[dict, None] = None,
+        request_id: str,
+        success: bool,
     ) -> Optional[float]:
         """Get resource usage cost associated with (current) synchronous processing request."""
 
-        user_id = user.user_id if user else user_id
+        user_id = user.user_id
 
         backend_config = get_backend_config()
 
@@ -1520,11 +1165,8 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
 
             etl_api = get_etl_api(
                 user=user,
-                allow_dynamic_etl_api=bool(
-                    # TODO #531 this is temporary feature flag, to removed when done
-                    backend_config.etl_dynamic_api_flag
-                    and flask.request.args.get(backend_config.etl_dynamic_api_flag)
-                ),
+                job_options=job_options,
+                allow_dynamic_etl_api=True,
                 requests_session=requests_session,
                 # TODO #531 provide a TtlCache here
                 etl_api_cache=None,
@@ -1532,7 +1174,7 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
 
             costs = etl_api.log_resource_usage(
                 batch_job_id=request_id,
-                title=None,
+                title=f"Synchronous processing request {request_id!r}",
                 execution_id=request_id,
                 user_id=user_id,
                 started_ms=None,
@@ -1548,8 +1190,7 @@ class GeoPySparkBackendImplementation(backend.OpenEoBackendImplementation):
             )
 
             logger.info(
-                f"{'successful' if success else 'failed'} request required {cpu_seconds} CPU-seconds, "
-                f"{mb_seconds} MB-seconds and {sentinel_hub_processing_units} PU(s); this cost {costs} credit(s)"
+                f"request_costs sync processing {request_id=} {success=} {cpu_seconds=} {mb_seconds=} {sentinel_hub_processing_units=} -> {costs=}"
             )
 
             return costs
@@ -1747,22 +1388,6 @@ def get_elastic_job_registry(
     return job_registry
 
 
-def extract_own_job_info(url: str, user_id: str, batch_jobs: backend.BatchJobs) -> Optional[BatchJobMetadata]:
-    path_segments = urlparse(url).path.split('/')
-
-    if len(path_segments) < 3:
-        return None
-
-    jobs_position_segment, job_id, results_position_segment = path_segments[-3:]
-    if jobs_position_segment != "jobs" or results_position_segment != "results":
-        return None
-
-    try:
-        return batch_jobs.get_job_info(job_id=job_id, user_id=user_id)
-    except JobNotFoundException:
-        return None
-
-
 class GpsBatchJobs(backend.BatchJobs):
 
     def __init__(
@@ -1901,7 +1526,7 @@ class GpsBatchJobs(backend.BatchJobs):
             """returns URL and (possibly empty) status for this job results dependency"""
             url = job_results_dependency['partial_job_results_url']
 
-            dependency_job_info = extract_own_job_info(url, user_id, batch_jobs=self)
+            dependency_job_info = load_stac.extract_own_job_info(url, user_id, batch_jobs=self)
             if dependency_job_info:
                 partial_job_status = PARTIAL_JOB_STATUS.for_job_status(dependency_job_info.status)
             else:
@@ -2248,19 +1873,19 @@ class GpsBatchJobs(backend.BatchJobs):
                                      status_code=400)
 
         isKube = ConfigParams().is_kube_deploy
-        driver_memory = job_options.get("driver-memory", "3G" if isKube else "8G" )
-        driver_memory_overhead = job_options.get("driver-memoryOverhead", "2G" if isKube else "2G")
-        executor_memory = job_options.get("executor-memory", "2G")
-        executor_memory_overhead = job_options.get("executor-memoryOverhead", "2500m" if isKube else "3G")
+        driver_memory = job_options.get("driver-memory", get_backend_config().default_driver_memory )
+        driver_memory_overhead = job_options.get("driver-memoryOverhead", get_backend_config().default_driver_memoryOverhead)
+        executor_memory = job_options.get("executor-memory", get_backend_config().default_executor_memory)
+        executor_memory_overhead = job_options.get("executor-memoryOverhead", get_backend_config().default_executor_memoryOverhead )
         driver_cores = str(job_options.get("driver-cores", 1 if isKube else 5))
         executor_cores = str(job_options.get("executor-cores", 1 if isKube else 2))
         executor_corerequest = job_options.get("executor-request-cores", "NONE")
         if executor_corerequest == "NONE":
             executor_corerequest = str(int(executor_cores)/2*1000)+"m"
-        max_executors = str(job_options.get("max-executors", 20 if isKube else 100))
-        executor_threads_jvm = str(job_options.get("executor-threads-jvm", 8 if isKube else 10))
-        gdal_dataset_cache_size = str(job_options.get("gdal-dataset-cache-size", 26))
-        gdal_cachemax = str(job_options.get("gdal-cachemax", 150))
+        max_executors = str(job_options.get("max-executors", get_backend_config().default_max_executors))
+        executor_threads_jvm = str(job_options.get("executor-threads-jvm", get_backend_config().default_executor_threads_jvm))
+        gdal_dataset_cache_size = str(job_options.get("gdal-dataset-cache-size", get_backend_config().default_gdal_dataset_cache_size))
+        gdal_cachemax = str(job_options.get("gdal-cachemax", get_backend_config().default_gdal_cachemax ))
         queue = job_options.get("queue", "default")
         profile = as_boolean_arg("profile", default_value="false")
         max_soft_errors_ratio = as_max_soft_errors_ratio_arg()
@@ -3012,7 +2637,7 @@ class GpsBatchJobs(backend.BatchJobs):
                 url, _ = arguments  # properties will be taken care of @ process graph evaluation time
 
                 if url.startswith("http://") or url.startswith("https://"):
-                    dependency_job_info = extract_own_job_info(url, user_id=user_id, batch_jobs=self)
+                    dependency_job_info = load_stac.extract_own_job_info(url, user_id=user_id, batch_jobs=self)
                     if dependency_job_info:
                         partial_job_status = PARTIAL_JOB_STATUS.for_job_status(dependency_job_info.status)
                     else: