+++ tests/test_backend.py
@@ -1,15 +1,14 @@
-import datetime as dt
 import mock
 import pytest
 import shapely
 
 from openeo_driver.ProcessGraphDeserializer import ENV_SOURCE_CONSTRAINTS
-from openeo_driver.backend import BatchJobs, BatchJobMetadata
 from openeo_driver.datacube import DriverVectorCube
 from openeo_driver.delayed_vector import DelayedVector
+from openeo_driver.users import User
 from openeo_driver.utils import EvalEnv
 
-from openeogeotrellis.backend import GpsBatchJobs, GpsProcessing, GeoPySparkBackendImplementation, extract_own_job_info
+from openeogeotrellis.backend import GpsBatchJobs, GpsProcessing, GeoPySparkBackendImplementation
 from openeogeotrellis.testing import gps_config_overrides
 
 
@@ -470,11 +469,13 @@ def test_request_costs(mock_get_etl_api_credentials_from_env, backend_implementa
         tracker = get_jvm.return_value.org.openeo.geotrelliscommon.ScopedMetadataTracker.apply.return_value
         tracker.sentinelHubProcessingUnits.return_value = shpu
 
-        credit_cost = backend_implementation.request_costs(user_id=user_id, request_id='r-abc123', success=success)
+        credit_cost = backend_implementation.request_costs(
+            user=User(user_id=user_id), request_id="r-abc123", success=success
+        )
 
         mock_etl_api.log_resource_usage.assert_called_once_with(
             batch_job_id='r-abc123',
-            title=None,
+            title="Synchronous processing request 'r-abc123'",
             execution_id='r-abc123',
             user_id=user_id,
             started_ms=None,
@@ -488,28 +489,3 @@ def test_request_costs(mock_get_etl_api_credentials_from_env, backend_implementa
         )
 
         assert credit_cost == 6
-
-
-@pytest.mark.parametrize("url, user_id, job_info_id",
-                         [
-                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results", 'alice', 'j-20240201abc123'),
-                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results", 'bob', None),
-                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results/N2Q1MjMzODEzNzRiNjJlNmYyYWFkMWYyZjlmYjZlZGRmNjI0ZDM4MmE4ZjcxZGI2Z/095be1c7a37baf63b2044?expires=1707382334", 'alice', None),
-                             ("https://oeo.net/openeo/1.1/jobs/j-20240201abc123/results/N2Q1MjMzODEzNzRiNjJlNmYyYWFkMWYyZjlmYjZlZGRmNjI0ZDM4MmE4ZjcxZGI2Z/095be1c7a37baf63b2044?expires=1707382334", 'bob', None),
-                             ("https://earth-search.aws.element84.com/v1/collections/sentinel-2-l2a", 'alice', None)
-                         ])
-def test_extract_own_job_info(url, user_id, job_info_id):
-    batch_jobs = mock.Mock(spec=BatchJobs)
-
-    def alices_single_job(job_id, user_id):
-        return (BatchJobMetadata(id=job_id, status='finished', created=dt.datetime.utcnow())
-                if job_id == 'j-20240201abc123' and user_id == 'alice' else None)
-
-    batch_jobs.get_job_info.side_effect = alices_single_job
-
-    job_info = extract_own_job_info(url, user_id, batch_jobs=batch_jobs)
-
-    if job_info_id is None:
-        assert job_info is None
-    else:
-        assert job_info.id == job_info_id