+++ CHANGELOG.md
@@ -12,6 +12,34 @@ without compromising stable operations.
 
 ## Unreleased
 
+## 0.32.0
+
+- Always enable `allow_dynamic_etl_api` from synchronous processing (drop feature flag)
+  ([#531](https://github.com/Open-EO/openeo-geopyspark-driver/issues/531), eu-cdse/openeo-cdse-infra#114)
+
+## 0.31.1
+
+- Initial support for `job_options` handling in `OpenEoBackendImplementation.request_costs()`
+  ([#531](https://github.com/Open-EO/openeo-geopyspark-driver/issues/531), eu-cdse/openeo-cdse-infra#114)
+
+## 0.31.0
+
+- vector_to_raster now returns a valid raster cube ([Open-EO/openeo-python-driver#273](https://github.com/Open-EO/openeo-python-driver/issues/273))
+- aggregate_spatial can now be used as input for vector_to_raster ([#663](https://github.com/Open-EO/openeo-geopyspark-driver/issues/663))
+- raster_to_vector now returns a valid vector cube ([Open-EO/openeo-python-driver#276](https://github.com/Open-EO/openeo-python-driver/issues/276))
+- raster_to_vector now includes the pixel values of the output polygons ([#578](https://github.com/Open-EO/openeo-geopyspark-driver/issues/578))
+- raster_to_vector now adds an id to each polygon including band name, date, and index ([#578](https://github.com/Open-EO/openeo-geopyspark-driver/issues/578))
+- resample_spatial now has better performance when resampling to a high resolution ([#265](https://github.com/Open-EO/openeo-geotrellis-extensions/issues/265))
+- support property filters for STAC based collections ([#460](https://github.com/Open-EO/openeo-geopyspark-driver/issues/460))
+- performance improvement when reading from object storage [#740](https://github.com/Open-EO/openeo-geopyspark-driver/issues/740)
+- support Microsoft Planetary Computer in load_stac [#760](https://github.com/Open-EO/openeo-geopyspark-driver/issues/760)
+
+## 0.30.2
+
+- load_stac: fix filtering by Item properties
+- Fix common cause of 'TopologyException' errors in vector processing
+- sample_by_feature will now use the 'feature_id_property' setting for naming generated assets. [#722](https://github.com/Open-EO/openeo-geopyspark-driver/issues/722)
+
 ## 0.30.1
 
 - Reinstate `ejr_credentials_vault_path` config option.
@@ -327,7 +355,7 @@ Build `20230117-966`
 - The 'if' process will no longer evaluate the branch that is not accepted https://github.com/Open-EO/openeo-python-driver/issues/109
 
 ## 2023-01-20 (0.6.7a1)
-- Changed: Getting a job's logs now leaves out log lines that have no loglevel or a level that is not supported. [openeo-python-driver/#160](https://github.com/Open-EO/openeo-python-driver/issues/160)
+- Changed: Getting a job's logs now leaves out log lines that have no loglevel or a level that is not supported. [Open-EO/openeo-python-driver#160](https://github.com/Open-EO/openeo-python-driver/issues/160)
 
 ## 2022-11-28 (0.6.3a1)
 - Added an experimental job option 'udf-dependency-archives' to pass on archives of UDF dependencies